Move marshalling functions in core Pipeline
